fixed issue with jti string conversion

+++ server/task/model.go
@@ -1,8 +1,8 @@
 package task
 
 type Task struct {
-	ID        uint   `json:"id"`
-	UserID    uint   `json:"user_id"`
+	ID        uint64 `json:"id"`
+	UserID    uint64 `json:"user_id"`
 	Title     string `json:"title" validate:"required"`
 	Completed bool   `json:"completed"`
 }
